Converted many constants to (small) tables, using alpha-indexed values from Roskam (0, 5, and 6.6).
Some minor rearrangement and renaming for consistency.

Experimentally added a complete loss of lift at 20 degrees alpha
(cancelling out CLo) for better stalling; we'll probably need to
fiddle with this number (it's just a wild guess, and might need to
come at a higher alpha angle).  It makes stall recovery much more
difficult, as it probably should be.