Rearrange the FC block such that:  * Elevator and rudder control are not affected by the PID controller  * Alpha limiter is not affected bij the PID controller  * yaw and pitch correction *are* affected by the PID controller
This also sets the G-limit to 9G better than before (it can get there
quicker and sustain longer).

There is also a problem that I don't yet have an answer to:

accelerations/n-pilot-z-norm has an offset of 1.0 when the aircraft is
horizontal and right side up. This needs to be corrected somehow but it
is not as simple as substracting 1.0 from it since the offset will be
different when the aircraft is pitching up (or rolled to one side).

there is probably need for one more property, or n-pilot-z-norm has to be fixed