Fixed the conversion factor inchtoft.
There are 12 inches in a foot so FGJSBBase::inchtoft should theoretically be equal to 1/12. However it has been historically set to 0.0833333 in FGJSBBase.cpp. This value is reasonably close to 1/12 and the approximation is completly harmless to JSBSim results.

The problem arises in automated tests where the conversion of inches to feet is made by dividing by 12. Unfortunately in some cases, the value to be checked is the result of a substraction between to "big" numbers that are almost equal. In that case, the difference between 1/12 and 0.08333333 might generate an error which is in the same order of magnitude than the expected result. A comparison where you check that the 2 values match by 7 figures might then fail.

The problem can be addressed in two ways:

1. Use the value 0.08333333 in the automated tests which makes the tests succeed (i.e. the compared values match to 7 figures). However if someone decide for any reason to increase the precision of FGJSBBase::inchtoft, then the tests might fail again. And the action of tracking the origin of the failure down to the difference between 0.08333333 and 1/12 is really a pain.

2. Use the theoretical value 1/12 in FGJSBBase. This does not make any noticeable difference to the computations made by JSBSim but it can then be safely assumed in automated tests that inches can be converted to feet by simply dividing by 12.

Hence this commit which modifies FGJSBBase::inchtoft from 0.08333333 to 1/12. This also allowed to increase the precision in the comparison made by some automated tests.