Added a means to convert units before exporting properties.
The unit conversion is made by defining a new type "meta" of <function> elements which use a dummy property #. This function is then applied to property values before exporting them.

The "meta" function is defined once and applied on as many properties as needed.
In addition, the user can create as many unit conversion functions as needed. These functions can be grouped in a separate file that can be included each time it is needed in a sort of library.

The new syntax is

<output>
  <function name="convertRadToDeg" type="meta">
    <product>
      <property> # </property>
      <value> 57.29577951 </value>
    </product>
  </function>
  <property caption="flight-path/gamma-deg" apply="convertRadToDeg">
    flight-path/gamma-rad
  </property>
</output>