Fixed an undesirable side effect in SetTemperature()
This bug was introduced by commit a4b78ea.

Before the commit above, FGStandardAtmosphere::SetTemperature() was setting TemperatureBias to 0.0 then was setting it to its final value. So at the first glance, it looked like the member TemperatureBias was set twice in a row. Hence the removal of the first occurence of `TemperatureBias=0.0` in the commit mentioned above.

Unfortunately, while updating the copy of JSBSim in FlightGear it appeared that this change broke SetTemperature(). Actually SetTemperature() was calling GetTemperature() to compute TemperatureBias but GetTemperature() was depending itself on TemperatureBias. Hence the need to reset TemperatureBias to 0.0 to break this circular dependency before calling GetTemperature().

The bug can be triggered by calling SetTemperature() twice in a row: after the 2nd call, the atmosphere temperature is incorrectly set.

This commit replaces the usage of GetTemperature() (and its vicious dependence on TemperatureBias) by an explicit computation which is based on the ISA temperature from GetStdTemperature() and corrected by the gradient TemperatureDeltaGradient.

Both the bug is squashed and the side effect removed.