Fixed the SEGFAULT triggered by the test of FG reset
This bug was introducedby by commit a4b78ea.

The bug is triggered by the reset sequence of FlightGear which does not delete the property manager when the instance of JSBSim is destroyed.

When a property is bound to the property manager there are basically 2 options:
1. The property does not yet exist, so it is created and the getter and setter from JSBSim are tied to the property.
2. The property already exists. In that case, the value currently stored in the property is passed to the setter that is tied by JSBSim.

However the properties binding takes place very early in the initialization of JSBSim (it is generally done by the class constructors). The problem happens when binding a setter to an existing property since, as it was mentioned above,  the current property value will be passed to the setter and if the class is improperly constructed, the setter will try to access data that is not yet available and will cause a SEGFAULT.

For that particular case, the SEGFAULT is happening when binding the property 'atmosphere/delta-T' during which the setter FGStandardAtmosphere::SetTemperatureBias() is called. During the execution of this setter, FGStandardAtmosphere::CalculatePressureBreakpoints() is called and tries to access the empty vector LapseRates which causes the SEGFAULT to be raised.

The bug is fixed by setting LapseRates to the required length. As a consequence, FGStandardAtmosphere::SetTemperatureBias() is no longer trying to access illegally the memory. Of course this call produces garbage since LapseRates is not yet initialized but that is not a problem since the call to FGStandardAtmosphere::InitModel() fixes that afterwards.