Removed file marking using the obsolete $Id$ tag.
This had been introduced at the time where CVS was the version control software used by JSBSim. Each file contained markings based on the CVS $Id$ tag. The purpose was to be able to determine which source files had been used to build a given JSBSim executable or library. Now that we are no longer using CVS, the $Id$ tags are no longer updated.

Nowadays we are using git which assigns a unique global ID (SHA) to each commit.This unique ID is now reported by the version string of JSBSim since commit cd884541c0. This made the file marking with $Id$ definitely obsolete.