<integrator> components are now internally replaced by <pid>
An <integrator> component is functionally the same thing than a <pid> with kp=0 and kd=0.0 and a trapezoidal integration scheme.

In the issue #122, it was requested to have the same features for the <integrator> component than the <pid> component. That would have resulted in duplicating some code from FGPID.cpp to FGFilter.cpp. In order to save the trouble of such a duplication, the <integrator> component is now internally modeled as a <pid> with the appropriates parameters.