Added the decorator @spare to ease debugging of test cases.
Whether a test succeeds or fails, all the temporary files are deleted at the end of the test by the method JSBSimTestCase.tearDown(). In order to keep some of these files for further examination a test method can decorated with @spare(filename) to spare the file named 'filename' from the sandbox deletion.