Fixed the fail_stuck behavior (bug reported in issue #119)
For sensors without a <lag> element, the output was set to zero when enabling `fail_stick` instead of keeping the last output from the sensor.

In addition, even for sensors with a <lag> element, the drift, gain, bias, delay and quantization were ignored since the output was stuck to the last output from Lag() therefore ignoring the other biases.