set-reset changed ...
I have analyzed the code to remove the flag "isSetProperty = true;" but I've seen that the method used: "setProperty = PropertyManager-> CreatePropertyObject <double> (element-> FindElementValue (" set "));" has a non-trivial problem. If you enter a property name in <reset> or <set> for example: "<reset> systems / gauges / PHI / doppler / test_reset_off </ reset>" if the property has not already been declared in the static list, the system returns a property warning not found, even if it has been defined in the code, but the worst thing is that running the value of the property is no longer read and therefore the behavior of the module becomes not in accordance with the wishes of those who wrote the code. Instead the method: "ptrReset = new FGParameterValue (property_string, PropertyManager);" it does not seem to have this problem. So I preferred to change <set> and <reset> with the same form used for <bias> and <versus>.