Add a new component FGLinearActuator (#160)
* New component LinearActuator
I have made a linear actuator to convert a rotation to several revolutions in a linear displacement. I discussed this device here in the "FlightGear ‹ Development ‹ Aircraft ‹ Flight dynamics model" and in GIT this topic. I installed a demo version on my local PC and I made a movie showing its features. I then discussed this object with @agodemar that advised me to proceed this way. For the work on the FIAT G91R1B sdi FGFS this type of actuator is very important and has the advantage of allowing very easily to obtain a correct rotation or movement of the instruments that perform multiple rotations, such as the compass, the radio-compass, the counters of miles with geneva gear etc. This object also introduces a set and reset command that makes it similar to a memory flip-flop with analogue output, an excellent system for creating a simple event counter. This object should not be confused with the currently present Actuator, as it is used before this device. The device has a system to discriminate the direction of rotation which can however be forced with an explicit direction of rotation. In this case, if the rotation direction does not coincide with the explicit direction, the device does not modify the output making it similar to an escapement. The instructions are in the file: models/flight_control/FGLinearActuator.h Obviously, as you have invited me to do @agodemar , I am willing to write a user manual with examples of this device and to maintain it. I will write a user manual with various examples of this device and to maintain it.

The testing of the module did not give me any problems, it was compiled directly for the development version of FGFS. Compared to the previous version, it has been introduced the possibility to vary the limitation in the direction of rotation (versus) and has been optimized. A bias has been introduced (which can be fixed or linked to a parameter) useful for managing tools that use offset commands. Honestly the name bias I used it as typical of an electronic control system, but for a linear actuator I think it is better to use the name "offset". In this phase of development it seems useful to me that others give advice in order to avoid the use of names not coherent with the function. There is also a hysteresis parameter (currently fixed) which allows to obtain an operation in steps typical of the step-motor or the escapement of the clocks. Something of the kind is also present in the classic actuator (which is often placed after this type of actuator), but this value operates on the input data and not in the output data. The hysteresis module allows the module to operate only if an input is large enough to exceed the hysteresis value. The use of hysteresis can lead to a considerable reduction in workload due to this module. An example is the use of the module to manage the gyrocompasses that have microvariations that can be blocked by the hysteresis without ever losing the consistency of the system. I think that another code of JSBSim could be reviewed in this logic avoiding an excessive use of the Execrete parameter.
For the output there is always the lag and the clip, theoretically these two parameters could be not inserted and left to a possible subsequent actuator module, but it is often convenient to have it and more compact the code. The code currently contains the bebug prints that can be activated / deactivated during the compilation of the code itself, will be removed in the final version, however some of these prints take place only in the case of effective mofification of the output.

* set-reset changed ...

I have analyzed the code to remove the flag "isSetProperty = true;" but I've seen that the method used: "setProperty = PropertyManager-> CreatePropertyObject <double> (element-> FindElementValue (" set "));" has a non-trivial problem. If you enter a property name in <reset> or <set> for example: "<reset> systems / gauges / PHI / doppler / test_reset_off </ reset>" if the property has not already been declared in the static list, the system returns a property warning not found, even if it has been defined in the code, but the worst thing is that running the value of the property is no longer read and therefore the behavior of the module becomes not in accordance with the wishes of those who wrote the code. Instead the method: "ptrReset = new FGParameterValue (property_string, PropertyManager);" it does not seem to have this problem. So I preferred to change <set> and <reset> with the same form used for <bias> and <versus>.