Fixed the bug reported in issue #123 (wrong HR errors at high alt).
The water vapor pressure is now capped to the maximum values observed on Earth according to ISA 1976 table 20 in order to avoid situations where the partial pressure of water vapor computed by the Magnus formula would exceed the ambient pressure.

The initialization of the lapse rates table & the pressure break points table had to be moved to the constructor of FGStandardAtmosphere in order to be able to call CalculatePressureAltitude() from routines tied to properties. Indeed when the property manager survives the FGFDMExec instance (as is tested by TestMiscellaneous.py) the routines that set atmosphere properties can be called before FGStandardAtmosphere::InitModel() is called and that was generating a SEGFAULT until the tables initializations have been moved to the constructor.

Would that occurence not been detected by JSBSim test suite, FlightGear would have crashed whenever someone would have tried to reset its flight.