Manage humidity with water to dry air mass ratio instead of the vapor partial pressure.
As is said in the ISA 1976 document: "relative humidity alone is not a physically meaningful indicator of atmospheric water vapor, humidity values have been reduced to mixing ratio (the mass of water vapor per unit mass of dry air)."

Following the same reasoning, FGStandardAtmosphere now manages water vapor as a water to dry air mass ratio.

The gas "constant" of the atmosphere is also updated according to the amount of water vapor in the air. This allowed to remove the specific method FGStandardAtmosphere::GetDensity() and to use FGAtmosphere::GetDensity() instead. As a consequence, the sound speed is now a function of both the ambient temperature and the relative humidity (RH). However the dependence to the RH is a second order factor.