Make sure nodes created by FCS functions are initialized.
The node creation from SGPropertyNode does not initialize the node value. For channels that are disabled at initialization this was causing a problem since the output nodes were left to their uninitialized state until the channel was activated. This was causing undesired side effects if some other systems were using these nodes in the meantime (bug reported by Torsten Renk).