Force the initialization of all members of FGPiston.
Investigations with Valgrind revealed that the class FGPiston had uninitialized members but it did not give details about exactly which were uninitialized.

This commit therefore uses the C++11 member initialization that allows to specify members initial values directly in the header. Rather through the tedious process of finding which member were initialized and which were not, I have specified initial values for every member of the class; even if it was redundant with the constructors initialization.