Relaxed the precision for some platforms.
Note: MinGW32 is twice less precise than other platforms for one computation. I don't want to relax the requirement for all platform just because one is performing below standards hence the use of #ifdef to keep track of that strange behavior. This needs more investigation: there is no reason for this, Linux uses gcc as well but gets better precision.