Fixed the compilation of Python packages (issue #285).
This is known to fail since compilation of Python packages with MinGW32 is longer supported by Python since version 3.5. These modifications are pushed to JSBSim for when the support of MinGW32 will be brought back by Python developers. Would that happen, this commit will hopefully avoid from delving again into the nuts and bolts of distutils.