Restoring the dependencies of jsbsim.pyx
The Python module was not rebuilt when the C++ headers which are included by jsbsim.pxd were modified. This was due to the INCLUDE_DIRECTORIES property not being passed to compile_pyx(). As a consequence, compile_pyx() was unable to build the path to the C++ headers and to check if they had been modified.

The root cause was a line in UseCython.cmake that I commented out more than 1 year ago. However, for some reason, the property INCLUDE_DIRECTORIES is not (and cannot be) set on the directory held by the CMAKE_CURRENT_BINARY_DIR variable. So the property is set on the source file jsbsim.pyx instead and compile_pyx() is instructed to find INCLUDE_DIRECTORIES from the source file jsbsim.pyx.