Use std::shared_ptr<> for FGPropertyManager
With this change, Python classes for JSBSim are now first class citizen regarding ownership/reference counting: a C++ class instance can no longer be destroyed when a Python object is referencing it.

The patch impacts a whole lot of files but it is just a trivial one-liner change for most of the files.