Fixed the dependency of the Python module to jsbsim.pxd
Declares jsbsim.pxd as a "configure file" so that CMake now knows that it must track its changes and update its copy in the build folder when required.

Failure to do so resulted in the jsbsim.pxd copy not being updated in the build folder and compile_pyx() not being called to rebuild the Python module when jsbsim.pxd was updated.