Fixed FGMagnetometer which updated the magnetic field every time step.
The counter was reset to zero each time the magnetic field was updated meaning that the update was processed again at the next time step or said otherwise the field was updated at each time step.

So removed the reset of the counter to zero (the variable 'counter' is unsigned int so we need more than 4 billions time step to overflow and, would that happen, the counter will be reset to zero which is completely harmless).

Also made the unit test more robust in the process.