Piston shall not send negative powers to thrusters
This happens when the engine is off: internal friction (static_friction_HP which is negative) is accounted for but the engine does not produce power so this internal friction is not counteracted resulting in an overall negative power. A better fix in the long term should therefore be to ignore static_friction_HP while the engine is not started.