Use templates for FGPropertyManager::Tie of a pointer.
According to the comments this method was instanciated manually for all types of pointers because otherwise it was confusing compilers. Well, it seems modern compilers can resolve the ambiguity so we can now use templates to avoid code duplication.

Took this opportunity to use nullptr instead of 0 in other templates.