Fix assertion failure following a FlightGear-like reset.
Bug fix by Coln Geniet:
For FGPropagate, tied properties are setup in the constructor by FGPropagate::bind. This happens before FGPropagate::InitModel is called, thus before vState.vLocation.setEllipse() is called. Hence when tied properties are setup, methods such as FGPropagate::SetAltitudeASL are not valid.

Now position/h-sl-ft happens to be tied with the option useDefault=true, so SetAltitudeASL will be called when tying it, if the property already exists (in which case, assert fails). When starting FG, this property does not exist, and nothing happens. However, changing location, JSBSim is reinitialized, but the property already exists, so the error occurs.

Changing useDefault to false in bind() fixes the problem.