Automatic update of the references to the stable release in README.md
Each time a new stable version is released, the corresponding references in README.md must be updated. Manual update is error prone not considering that it might be overlooked by maintainers until someone opens an issue (see issue #337 and PR #338).

This commit fixes that and automatically pushes changes to README.md to update all the references when a new stable release is built.