Filters  out negative powers issued to the propeller.
As reported by FlightGear ticket 2581 as well as in its mailing list, some issues have araised following the commit 22f8330 (applying the engine torque to the vehicle).

Most engines are issuing a negative power when turned off in order to slow the propeller down "more realistically". The problem is that this negative power was unconditionally produced and it resulted in an infinite torque when the propeller was motionless (RPM==0.0) since Torque=Power/RPM.

The purpose of this commit is therefore to make sure that none of the electric, piston and turboprop engines are sending a negative power when the propeller is motionless (i.e. as RPM <= 0.1).