Fixed property "aero/h_b-mac-ft" algorithm error
The algorithm of property "aero/h_b-mac-ft" in JSBSim is wrong. This error will lead to an error in the calculation of aircraft ground effect. In fact, the Z axis of the body frame is downward. Assuming that the aircraft stops on the ground, when MAC is higher than the CG, vMac(3) is negative, resulting in hoverbmac less than hoverbcg. In fact, hoverbmac should be greater than hoverbcg. Therefore, vMac(3) should be inverted