Fixed a sign error in the computation of "aero/h_b-mac-ft" (#508)
The AeroRP is translated from the body frame of reference to the local (NED) frame of reference, via Tb2l, which also has it's Down axis as positive pointing down. Therefore, the sign of vMac(3) should be inverted. This error will lead to an error in the calculation of aircraft ground effect.