Prepare the obsolescence of Python 3.6
Security fixes for Python 3.6 will no longer be released in 2022. JSBSim will then drop the support of Python 3.6.

Setting the `python_requires` parameter in `setup.py` allows PyPI determining which is the last JSBSim release that is compatible with Python 3.6. In addition, this will allow future JSBSim releases to use features from Python 3.7 that are not compatible with 3.6.