Make `set_root_dir` compliant with the C++ method.
Unlike the C++ counterpart, the Python method `set_root_dir` was updating the aircraft, engine and systems path. This is no longer the case and `set_root_dir` now complies with the behavior of the C++ method `FGFDMExec::SetRootDir`.

However the Python constructor of `FGFDMExec` still updates the aircraft, engine and systems paths to keep backward compatibility i.e. being able to set up `FGFDMExec` with only one line of code.

Beyond that, `set_aircraft_path` and the likes must be called after `set_root_dir` if the paths need to be updated. Just as for the C++ method.