Prepare the migration from the `setuptools` frontend to the `build` module.
Calls to `setup.py` from the command line will become obsolete at some point. Python 3.10 and recent versions of `setuptools` are already displaying warnings about using `python setup.py install`.
This commit initiates the migration of the CI workflow and of the shell script `build-wheels.sh` to the `build` module in compliance with PEP517 and PEP518.