Don't ignore files which name starts with `src/JSBSim`
There are already several source files matching this criteria so there is no reason to ignore them.

The pattern was most likely meant to ignore the executables `JSBSim` (Linux) and `JSBSim.exe`  (Windows) but this is now irrelevant as JSBSim uses CMake which builds the project in a separate folder.

For the sake of completeness, this is replaced by a pattern that ignores executables i.e. `*.exe` in Windows parlance.