Build JSBSim as a DLL (#644)
JSBSim can now be compiled as a DLL as requested in discussion #631.

* The flag `JSBSIM_EXPORT` must be defined when compiling JSBSim as a DLL. It then must be undefined when linking an executable with the DLL of JSBSim.
* The flag `JSBSIM_STATIC_LINK` must be defined when compiling **and** linking with JSBSim as a static library.

CMake takes care of defining the correct flags depending on the CMake flag `BUILD_SHARED_LIBS` being set to `ON` or `OFF`. Default is `BUILD_SHARED_LIBS=OFF` i.e. CMake builds the static library by default.

The GitHub Actions workflow `Windows-MSVC` is modified to test compile JSBSim with MSVC for both the static library and the dynamic library.