Update `libexpat` to the latest version 2.4.8
The `libexpat` code included in JSBSim is now pristine from the `libexpat` repo. No hack whatsoever has been made. This will help maintaining the `libexpat` code in JSBSim to the most recent updates, including security fixes.

The `libexpat` license is now included in our repo and copied in our distribution packages such as the Windows installer and the Python module.