Remove the dependency to `win32api`.
The package `pywin32` from which the `win32api` module is coming is not portable as it is platform specific (Windows only). As a consequence, it cannot be managed by the build tools such as `pip wheel` or `build` unless relying on complex, fragile, hard to maintain code.

Actually the `win32api` features that were used happen to be irrelevant for the problem that our code is addressing as the path to the Windows folder `C:\Windows` and to the Windows system folder `C:\Windows\System32` are included by default in the environment variable `$PATH`. So we don't need `win32api` to include them in the list of folders to visit.