Handles the case where the documentation should be built from a JSBSim module installed with `pip`.
If `cmake` is called but the module `_jsbsim` is not built, then Sphinx complains because the build directory contains an incomplete `jsbsim` module.

This commit works around that situation by allowing Sphinx to use a version of JSBSim that has been globally installed (with `pip` for instance) rather than the half-baked version that may reside in `build/tests`.

The purpose of this commit is to allow building the Python docs with a JSBSim module that has been installed by `pip`. This saves an additional complete build of the Python module when releasing JSBSim with our CI/CD workflow in GitHub.